requirements.txt
langchain>=0.3.13
langchain-community>=0.3.13
requests>=2.31.0
pydantic>=2.5.2
PyYAML>=6.0.1
python-dotenv>=1.0.0
openai>=1.3.0
pytest>=7.4.3
pytest-mock>=3.12.0
requests-mock>=1.11.0